implemented on-off switch for detectors
