+++ examples/optimize_ceviche_corner.py
@@ -24,7 +24,7 @@ from fdtdx.objects.boundaries import BoundaryConfig, boundary_objects_from_confi
 from fdtdx.objects.detectors import EnergyDetector, PoyntingFluxDetector
 from fdtdx.objects.device.parameters.mapping import DiscreteParameterMapping
 from fdtdx.objects.sources import ModePlaneSource
-from fdtdx.core import metric_efficiency
+from fdtdx.core import metric_efficiency, OnOffSwitch
 from fdtdx.utils import Logger, plot_setup
 
 
@@ -163,7 +163,7 @@ def main(
         name="Input Detector",
         partial_grid_shape=(1, None, None),
         direction="+",
-        time_steps=all_time_steps[3 * period_steps : 5 * period_steps],
+        switch=OnOffSwitch(fixed_on_time_steps=all_time_steps[3 * period_steps : 5 * period_steps]),
     )
     placement_constraints.extend(
         [
@@ -205,7 +205,7 @@ def main(
         name="Output Detector",
         partial_grid_shape=(None, 1, None),
         direction="+",
-        time_steps=all_time_steps[-2 * period_steps :],
+        switch=OnOffSwitch(fixed_on_time_steps=all_time_steps[-2 * period_steps :]),
     )
     placement_constraints.extend(
         [
@@ -226,7 +226,7 @@ def main(
     energy_last_step = EnergyDetector(
         name="energy_last_step",
         as_slices=True,
-        time_steps=[-1],
+        switch=OnOffSwitch(fixed_on_time_steps=[-1]),
     )
     placement_constraints.extend([*energy_last_step.same_position_and_size(volume)])
 
@@ -235,7 +235,7 @@ def main(
         video_detector = EnergyDetector(
             name="video",
             as_slices=True,
-            interval=10,
+            switch=OnOffSwitch(interval=10),
             exact_interpolation=True,
         )
         placement_constraints.extend([*video_detector.same_position_and_size(volume)])
@@ -245,7 +245,7 @@ def main(
                 name="backward_video",
                 as_slices=True,
                 inverse=True,
-                interval=10,
+                switch=OnOffSwitch(interval=10),
                 exact_interpolation=True,
             )
             placement_constraints.extend([*backward_video_detector.same_position_and_size(volume)])
