+++ src/fdtdx/core/misc.py
@@ -11,117 +11,6 @@ from fdtdx.core.linalg import get_orthogonal_vector
 from fdtdx.materials import Material
 
 
-def is_on_at_time_step(
-    is_on: bool,
-    start_time: float | None,
-    start_after_periods: float | None,
-    end_time: float | None,
-    end_after_periods: float | None,
-    on_for_time: float | None,
-    on_for_periods: float | None,
-    time_step: int,
-    time_step_duration: float,
-    period: float | None,
-) -> bool:  # scalar bool
-    """Determines if a time-dependent component should be active at a given time step.
-
-    Args:
-        is_on: Base on/off state
-        start_time: Absolute start time
-        start_after_periods: Start time in terms of periods
-        end_time: Absolute end time
-        end_after_periods: End time in terms of periods
-        on_for_time: Duration to stay on in absolute time
-        on_for_periods: Duration to stay on in terms of periods
-        time_step: Current simulation time step
-        time_step_duration: Duration of each time step
-        period: Period length for period-based timing
-
-    Returns:
-        bool: True if the component should be active at the given time step
-    """
-    if not is_on:
-        return False
-
-    # validate start/end/on time
-    if (
-        any(
-            [
-                x is not None
-                for x in [
-                    start_after_periods,
-                    end_after_periods,
-                    on_for_periods,
-                ]
-            ]
-        )
-        and period is None
-    ):
-        raise Exception("Need to specify period!")
-    num_start_specs = sum(
-        [
-            start_time is not None,
-            start_after_periods is not None,
-            on_for_time is not None and end_time is not None,
-            on_for_periods is not None and end_time is not None,
-            on_for_time is not None and end_after_periods is not None,
-            on_for_periods is not None and end_after_periods is not None,
-        ]
-    )
-    if num_start_specs > 1:
-        raise Exception("Invalid start time specification!")
-    if num_start_specs == 0:
-        start_time = 0
-    num_end_specs = sum(
-        [
-            end_time is not None,
-            end_after_periods is not None,
-            on_for_time is not None and start_time is not None,
-            on_for_periods is not None and start_time is not None,
-            on_for_time is not None and start_after_periods is not None,
-            on_for_periods is not None and start_after_periods is not None,
-        ]
-    )
-    if num_end_specs > 1:
-        raise Exception("Invalid end time specification!")
-    if num_end_specs == 0:
-        end_time = math.inf
-
-    # period to actual time
-    if start_after_periods is not None:
-        if period is None:
-            raise Exception("This should never happen")
-        start_time = start_after_periods * period
-    if end_after_periods is not None:
-        if period is None:
-            raise Exception("This should never happen")
-        end_time = end_after_periods * period
-    if on_for_periods is not None:
-        if period is None:
-            raise Exception("This should never happen")
-        on_for_time = on_for_periods * period
-
-    # determine start/end time
-    if start_time is None and on_for_time is not None:
-        if end_time is None:
-            raise Exception("This should never happen")
-        start_time = end_time - on_for_time
-
-    if end_time is None and on_for_time is not None:
-        if start_time is None:
-            raise Exception("This should never happen")
-        end_time = start_time + on_for_time
-
-    # check if on
-    if start_time is None or end_time is None:
-        raise Exception("This should never happen")
-    time_passed = time_step * time_step_duration
-    on = True
-    on = on and (start_time <= time_passed)
-    on = on and (time_passed <= end_time)
-    return on
-
-
 def expand_matrix(matrix: jax.Array, grid_points_per_voxel: tuple[int, ...]) -> jax.Array:
     """Expands a matrix by repeating values along spatial dimensions and optionally adding channels.
